fix: close production-readiness gaps across CI gates, pipeline, contracts, and features
Multi-surface remediation that brings the deploy pipeline green from a clean
checkout and fixes real runtime defects across the codebase.

CI gates (were red on ci.yml + release.yml, now green):
- mypy: enable the pydantic.mypy plugin + pin a compatible mypy/pydantic
  pair, and run with --config-file cortex/pyproject.toml so the per-module
  overrides actually apply (the config was undiscoverable from the repo
  root -> 4956 errors). Tree driven to zero under --strict.
- pytest: replace the single segfaulting `pytest cortex/tests/` with a
  2-invocation split (whole suite except the PySide6 desktop_shell test,
  then that test isolated). Restores integration/state_engine/eval/physio/
  performance coverage the earlier 3-way split silently dropped.
- extension tsc: fix all 37 errors (chrome.tabGroups.Color, non-empty tab
  tuples, typed WS payloads / storage / LeetCode session shapes).
- vscode tsc: fix the phase union + String-not-callable so the VSIX builds;
  add a vscode tsc CI job.

Pipeline / backend correctness:
- physio: gate respiration_rate on validity (was raising ValueError that
  silently dropped whole physio+kinematics frames); thread the real
  blink_suppression score + frame timestamp so screen-apnea can fire; live
  motion-respiration fusion; documented ear/nose forward-lean geometry;
  mark replayed face frames; default face_mesh_subsample_n=1.
- store: RedisConfig.enabled defaults False and the in-memory fallback gets
  a persist_path, so consent/calibration state survives restarts.
- state/intervention: wall-clock weekly-schedule gate; single dismissal
  increment; FLOW warm-up gate; consent recorded under the canonical
  action-types the executor gates on (escalation now works); copilot
  throttle force-enabled on shutdown.
- gateway: /api/cost reports the real configured budget; /llm/plan reports
  fallback_used; STATE_UPDATE degraded/source derived from signal quality.
- observability: configure_logging wired at every entrypoint; Prometheus
  counters and structured events emitted at their real callsites.

Contracts (schema == wire; TS regenerated):
- capability token attached to /api/cost + /api/feedback; user_agent on
  FeedbackRequest; InterventionPlan.trigger_url stamped and used for the
  per-site dismissal cooldown; SessionRecap wrapper; phase 'execute_action'.

Desktop / extension:
- controller routes native actions natively (biology break, resume file);
  WS-mode calibration pauses the daemon and surfaces a visible error
  instead of silently calibrating on synthetic frames; activity-tracker
  registered as a Plasmo content script (was never injected); honest
  vibrancy / onboarding-notification / connections behaviour.

Tests: add coverage for every fix and repair order-dependent asyncio
event-loop helpers so the full suite passes in a single process.

+++ cortex/services/state_engine/leetcode_mode_resolver.py
@@ -202,10 +202,10 @@ class LeetCodeModeResolver:
         if mode == LeetCodeMode.AMYGDALA_HIJACK:
             # Use aai_score itself as a proxy for detector confidence
             detector_conf = min(aai_score, 1.0)
-            return (state_confidence * detector_conf) ** 0.5
+            return float((state_confidence * detector_conf) ** 0.5)
 
         if mode == LeetCodeMode.DESTRUCTIVE_STRUGGLE:
-            return (state_confidence * destructive.confidence) ** 0.5
+            return float((state_confidence * destructive.confidence) ** 0.5)
 
         if mode == LeetCodeMode.FATIGUE:
             # No separate detector confidence; use stress as proxy
